requirements.txt

streamlit==1.37.1
pdfminer.six==20231228
python-docx==1.1.2
